Refactored events filter reducer
