+++ EventsExpress/ClientApp/src/reducers/events-filter/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import { organizerFilterReducer } from './organizer-filter';
+
+export const reducer = combineReducers({
+    organizerFilter: organizerFilterReducer
+});
